backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0
dj-database-url==2.1.0
# ML and Audio Processing
scikit-learn==1.3.2
librosa==0.10.1
numpy==1.24.3
pandas==2.0.3
joblib==1.3.2
# File handling
Pillow==10.1.0
# API and WebSocket support
channels==4.0.0
channels-redis==4.1.0
celery==5.3.4
redis==5.0.1
